start cleanup for testnet release

+++ functions/initMQ.js
@@ -1,12 +1,14 @@
 const RSMQPromise = require('rsmq-promise');
 
-async function initMQ(redis, eventNames) {
+async function initMQ(redis, eventNames, purge) {
     const mq = new RSMQPromise(redis);
     let allQueues = await mq.listQueues();
-    for (const queue of allQueues) {
-        await mq.deleteQueue({qname: queue});
+    if (purge !== undefined && purge === true) {
+        for (const queue of allQueues) {
+            await mq.deleteQueue({qname: queue});
+        }
+        allQueues = await mq.listQueues();
     }
-    allQueues = await mq.listQueues();
 	for (const eventName of eventNames) {
 	    const qname = `event-${eventName}`;
 	    if (!allQueues.includes(qname)) {
